Phyics.MultipletCombining

Description

This module contains operators and functions for combining SU(n) multiplets according to the algorithm presented by C.G. Wohl in the PDG book 2021 section 48 https://pdg.lbl.gov/2022/reviews/rpp2022-rev-young-diagrams.pdf.

It provides the operators (><) and (>><) for combining multiplets, and the function multi and multis to calculate the multiplicities, e.g.:

    [1,0] >< [0,1] = [[1,1],[0,0]]

    multi [1,0] = 3

    [1,0] >< [1,0] >>< [1,0] = [[3,0],[1,1],[1,1],[0,0]]

    multis $ [1,0] >< [1,0] >>< [1,0] = [10,8,8,1]

Example for combinaing two multiplets using Young-Diagrams:

    (0,0)x(0,0) = (0,0)
    #     a   # a  # a   # a
    # (x) b = #  > # b > # b
    #     c   #    #     # c

    (1,0)x(1,0) = (step ->)   (2,0)      +   (step ->) (0,1)
    # #    a a    # # a a    # # a a           # # a    # # a
    #   x  b    = #       >  # b         +     # a    > # a b
    #      c      #          # c               #        # c

Kroneker product like operators

(><) :: [Int] -> [Int] -> [[Int]]

Produce multiplet structure from combining two SU(n) multiplets

(>><) :: [[Int]] -> [Int] -> [[Int]]

Produce multiplet structure from combining a list of multiplets with another multiplet

Multiplicitiy calculation

multi :: [Int] -> Int

Calculate the multiplicity of a multiplet

multis :: [[Int]] -> [Int]

Calculate the multiplicities of a list of multiplets

Basic data type

data Tableau

Basic type used for a Tableau/Diagram

Lower level functions

yt_symbols :: [Int] -> Tableau

Build a tableau bottom up from it's label.

yts_symbols :: [[Int]] -> [Tableau]

Build multiple tableaux from multiple labels.

showt :: [Tableau] -> String

Show like function to display a list of tableaux.

yt_nums :: Tableau -> [Int]

Calculate the number representation from a tableau.

yts_nums :: [Tableau] -> [[Int]]

Calculate the list of labels fro a list of tableaux.

admis :: String -> Bool

Check for the string for being composed of admissible letters. | Admissible and not admissible examples:

admis "aabacdaebbcbd"  = True

last letter not admissable
admis "abacae"  = False
admis "abacdec"  = False

unchain :: Char -> String -> Maybe String

Extract one strictly ordered chain from the given string, starting at the given character.

sym2letter :: Tableau -> Tableau

Convert a tableau of symbols into a tableau of letters

appendAt :: Int -> String -> Tableau -> Tableau

Append a string to the i'th line of a tableau.

readTab :: Tableau -> String

Read a string of letters from a given tableau to be checked for admissibility.

combis :: Int -> Int -> [[Int]]

Produce a list of placing-coordinates of all combinations for a tableau with t rows to place c character.

E.g.: 3 rows, two characters -> 3*3 possible placements:

1:1, 1:2, 1:3, 2:1, 2:2, 2:3, 3:1, 3:2, 3:3

tabs1 :: Tableau -> String -> [Tableau]

Create multiple new tableau using newtab given one tableau and one line of a right side tableau.

e.g.: tabs1 (yt_symbols [1,1,1]) "a a "

allTsFromSyms :: Tableau -> Tableau -> [Tableau]

Create all tableaux from two given tableaux.

allTs :: [Int] -> [Int] -> [Tableau]

allTs [1,0] [1,1]

Create all tableau from two tableaux identified by their labels.

  putStrLn $ showt $ noDoubs.admisTabs $ allTs [1,1] [1,1]
  yts_nums $ noDoubs.admisTabs $ allTs [1,1] [1,1]
  [[2,2],[3,0],[0,3],[1,1],[1,1],[0,0]]